Various gallery changes
Changed gallery in order to achieve mobile responsiveness. Aria labels added to buttons in order to improve accessibility. Introduction of image modal upon clicking an image in the image gallery. Coded for desktop and mobile. Images to be chosen with client and further work with image modal carousel to be carried out.